# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Fem-ho
  version: 0.1.0
  description: |
    Contracte de l'API de Fem-ho. Aquest fitxer és la font de veritat (regla 5
    d'instruccions.md): els tipus de TypeScript i el client de Kotlin es generen des
    d'aquí i cap client escriu tipus a mà. Un endpoint que no hi és no existeix.

    A M1 només hi ha les dues rutes públiques de vida i identitat. La resta arriba a
    la seva fita, seguint docs/05-api-rest.md.
  license:
    name: AGPL-3.0-or-later
    identifier: AGPL-3.0-or-later
servers:
  - url: /api/v1
    description: Base de l'API. GET /info i GET /healthz pengen de l'arrel, no d'aquí.
...
paths:
...
  /tokens:
    get:
      tags: [tokens]
      operationId: listApiTokens
      summary: Els tokens d'API de qui pregunta
      description: |
        Es llisten amb el **prefix** i l'últim ús, mai el token sencer: només se'n guarda
        el hash (docs/05 §2). Si l'usuari el perd, ha de crear-ne un de nou.
      responses:
        '200':
          description: Els tokens.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/ApiTokenSummary' }
        '401':
          $ref: '#/components/responses/Unauthenticated'
    post:
      tags: [tokens]
      operationId: createApiToken
      summary: Crear un token d'API
      description: |
        **El token sencer es retorna UN SOL COP.** No es pot recuperar del hash, i cal
        dir-ho clarament a qui el crea.

        Les capacitats i els àmbits van al registre del token, **no** a scopes d'OAuth
        (regla 9).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, capabilities]
              properties:
                name: { type: string }
                capabilities:
                  type: array
                  items: { type: string }
                scope_ids:
                  type: array
                  items: { type: string }
                  description: Buit vol dir tots els àmbits del propietari.
                expires_at: { type: [string, 'null'], format: date-time }
      responses:
        '201':
          description: Creat. Porta el token en clar, i és l'única vegada.
          content:
            application/json:
              schema:
                type: object
                required: [token, summary]
                properties:
                  token:
                    type: string
                    description: El token sencer. No es torna a poder llegir mai més.
                  summary: { $ref: '#/components/schemas/ApiTokenSummary' }
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
  /tokens/{id}:
    delete:
      tags: [tokens]
      operationId: revokeApiToken
      summary: Revocar un token
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '204':
          description: Revocat.
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /tasks/{id}/ai-mode:
    post:
      tags: [tasks]
      operationId: setAiMode
      summary: Canviar el mode d'IA d'una tasca
      description: |
        Els tres modes de docs/09 §2. `manual` és el per defecte i no es pinta res a la
        targeta; `assisted` pot llegir, comentar i preparar però **no completar**;
        `delegated` la pot executar sencera.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ai_mode]
              properties:
                ai_mode:
                  type: string
                  enum: [manual, assisted, delegated]
      responses:
        '200':
          description: El mode nou.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /tasks/{id}/activity:
    get:
      tags: [tasks]
      operationId: listActivity
      summary: L'historial d'una tasca
      description: |
        Els actors barrejats —humans, IA i externs— amb el valor anterior i el nou de
        cada camp (docs/09 §7).
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: actor
          in: query
          required: false
          schema: { type: string, enum: [all, ai, human] }
      responses:
        '200':
          description: L'historial.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/ActivityEntry' }
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /tasks/{id}:
    get:
      tags: [tasks]
      operationId: getTask
      summary: Una tasca
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200':
          description: La tasca.
          content: { application/json: { schema: { $ref: '#/components/schemas/Task' } } }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [tasks]
      operationId: updateTask
      summary: Modificar una tasca
      description: |
        Només els camps que es donin. `null` vol dir "buida'l" i absent vol dir "no el
        toquis": sense la distinció, buidar una data seria impossible des d'un client
        que envia només el que ha canviat.
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/TaskPatch' } } }
      responses:
        '200':
          description: Modificada.
          content: { application/json: { schema: { $ref: '#/components/schemas/Task' } } }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [tasks]
      operationId: deleteTask
      summary: Esborrar una tasca
      description: |
        Esborrat suau. Les subtasques i les llistes cauen amb ella —no existeixen fora
        d'una tasca—, a diferència del que passa en esborrar un projecte.
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '204': { description: Esborrada. }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  schemas:
...
    TaskPatch:
      type: object
      description: |
        `null` vol dir "buida'l" i absent vol dir "no el toquis". Per moure-la, `/move`.
      properties:
        title: { type: string }
        description: { type: [string, 'null'] }
        due_date: { type: [string, 'null'], format: date }
        due_time: { type: [string, 'null'] }
        deadline:
          type: [string, 'null']
          format: date-time
          description: |
            Separada del venciment (docs/02 §7): "fes-ho aquest dijous" i "com a molt
            tard el dia 30" són dues coses, i amb un sol camp qui té les dues n'ha de
            triar una.
        project_id:
          type: [string, 'null']
          description: |
            `null` la torna a l'espai general de l'àmbit. **L'àmbit no es pot canviar
            editant**: altres membres, altres etiquetes i l'assignació automàtica dels
            àmbits individuals. Un projecte d'un altre àmbit es rebutja amb `422`.
        rrule:
          type: [string, 'null']
          description: 'RFC 5545. `null` deixa de repetir-se.'
        recurrence_mode:
          type: string
          enum: [schedule, completion]
          description: |
            La distinció que Todoist escriu com a `every` contra `every!`: repetir-se
            cada dimarts contra repetir-se una setmana **després d'haver-la fet**. RRULE
            no sap expressar la segona, i per a tasques domèstiques és la que la gent
            vol; per això és un camp propi (docs/01 §4).
        ai_mode: { type: string, enum: [manual, assisted, delegated] }
        ai_instructions: { type: [string, 'null'] }
...
    ApiTokenSummary:
      type: object
      required: [id, name, token_prefix, capabilities, scope_ids, created_at]
      properties:
        id: { type: string }
        name: { type: string }
        token_prefix:
          type: string
          description: Els primers caràcters, per poder-lo reconèixer a la llista.
        capabilities:
          type: array
          items: { type: string }
        scope_ids:
          type: array
          items: { type: string }
        created_at: { type: string, format: date-time }
        expires_at: { type: [string, 'null'], format: date-time }
        last_used_at: { type: [string, 'null'], format: date-time }
        revoked_at: { type: [string, 'null'], format: date-time }
    ActivityEntry:
      type: object
      required: [id, entity_type, entity_id, verb, actor_type, source, created_at]
      properties:
        id:
          type: string
          description: UUIDv7, que ordena per temps sense un comptador a part.
        entity_type: { type: string }
        entity_id: { type: string }
        verb: { type: string }
        actor_type:
          type: string
          enum: [user, ai_agent, guest, system, caldav]
        actor_user_id: { type: [string, 'null'] }
        actor_agent_id: { type: [string, 'null'] }
        actor_label:
          type: [string, 'null']
          description: '"Borja", "IA · Claude", "Extern · Marta".'
        source: { type: string }
        changes:
          type: [object, 'null']
          description: |
            `{camp: {from, to}}`. És el que fa possible el botó Desfés: sense el valor
            anterior no hi ha res a què tornar.

            La forma és fixa i el contracte ho diu, perquè `additionalProperties: true`
            generava un `unknown` per valor i obligava cada client a assumir-ne la forma
            pel seu compte, que és exactament el que la regla 5 evita.
          additionalProperties:
            type: object
            required: [from, to]
            properties:
              from: {}
              to: {}
        created_at: { type: string, format: date-time }
        undoable:
          type: boolean
          description: Un canvi autònom d'IA que encara es pot desfer.
...
    Task:
      type: object
      additionalProperties: false
      required:
        [id, scope_id, title, status, position, ai_mode, view_mode, created_at, updated_at, version]
      properties:
        id: { type: string }
        scope_id: { type: string }
        project_id: { type: [string, 'null'] }
        title: { type: string }
        description: { type: [string, 'null'] }
        status:
          type: string
          enum: [inbox, todo, doing, done]
          description: La columna. Es diu `status`, no `column` (D2).
        position:
          type: string
          description: Índex fraccional. Ordenat amb collation binària (D3).
        due_date:
          type: [string, 'null']
          description: YYYY-MM-DD, sense fus. Una data de tot el dia no té instant.
        due_time: { type: [string, 'null'] }
        deadline: { type: [string, 'null'], format: date-time }
        completed_at: { type: [string, 'null'], format: date-time }
        view_mode:
          type: string
          enum: [card, simple]
        ai_mode:
          type: string
          enum: [manual, assisted, delegated]
          description: Delegar no és assignar (D5).
        delegate_agent_id: { type: [string, 'null'] }
        rrule:
          type: [string, 'null']
          description: 'RFC 5545, o `null` si no es repeteix.'
        recurrence_mode:
          type: [string, 'null']
          enum: [schedule, completion, null]
          description: |
            `schedule` compta des del venciment; `completion`, des que es fa. És la
            distinció que Todoist escriu com a `every` contra `every!` i que RRULE no sap
            expressar (docs/01 §4).
        recurrence_parent_id:
          type: [string, 'null']
          description: La instància de la qual ve, si n'és una repetició.
        progress:
          type: object
          required: [done, total, lists]
          description: |
            Subtasques i ítems de llista **comptats junts**. La targeta els ensenya sota
            un sol commutador ("3/7 fets") perquè per a qui mira el tauler són el mateix:
            coses que falten dins d'aquesta tasca. La distinció importa en obrir-la.

            És un agregat i no la llista: portar tots els ítems de totes les targetes
            faria baixar milers de files per pintar unes quantes pastilles.
          properties:
            done: { type: integer }
            total: { type: integer }
            lists:
              type: integer
              description: |
                Quants blocs desplegables té la targeta: un per les subtasques, si en té,
                i un per cada llista. És el número del commutador "Llistes (2)", que
                compta blocs i no ítems.
        assignee_ids:
          type: array
          items: { type: string }
          description: Persones. El brief demana "persona o persones".
        created_by: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        version: { type: integer }
...
    Problem:
      type: object
      description: |
        Error en application/problem+json (RFC 9457, docs/05 §3).

        **Qui llegeix un error no és qui el pateix.** El reben tres públics molt
        diferents: una persona davant de l'app, un agent per MCP, i un client CalDAV com
        DAVx⁵ o Thunderbird. Els dos últims no tenen ni catàleg ni idioma. Per això
        l'error viatja en dues peces:

        - `type`, `title` i `detail` **en anglès**, estables, per programar-hi.
        - `params`, les **dades** de l'error. Les apps hi posen el text del seu catàleg
          amb la clau `error.<slug>`, on el `slug` és l'últim tram del `type`. Si no en
          tenen la clau, ensenyen el `detail`: així un error nou del servidor no deixa
          mai una pantalla muda ni obliga a desplegar les dues bandes alhora.
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          examples: ['https://femho.app/errors/scope-forbidden']
        title:
          type: string
          examples: ['Scope not accessible']
        params:
          type: object
          additionalProperties: { type: [string, integer] }
          description: |
            Les dades de l'error, planes. Plans a posta: el client no ha de saber
            recórrer un JSON dins d'un JSON per pintar un missatge.
          examples: [{ name: 'Família', count: 3 }]
        status:
          type: integer
          examples: [403]
        detail:
          type: string
          examples: ["Aquest token només té accés a l'àmbit Feina."]
        instance:
          type: string
  responses:
    Unauthenticated:
      description: Cal autenticar-se.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: |
        Sense permís. El cos diu QUINS àmbits veu el token i ON és el que s'ha demanat:
        un 403 mut fa que un agent reintenti en bucle (docs/05 §2).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: No existeix, o no és visible per a qui pregunta.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: femho_session
      description: Sessió de navegador. Galeta HttpOnly, Secure, SameSite=Lax.
    bearer:
      type: http
      scheme: bearer
      description: |
        Token d'accés (Android) o token d'API `femho_pat_…` (integracions, MCP, IA).
        Les capacitats i els àmbits van al registre del token, no a scopes d'OAuth
        (regla 9).
security:
  - session: []
  - bearer: []